Move global constants into globals.py
This allows them to be imported into any other files as required.
The alternative was to pass the required values as arguments to
functions and class constructors, but that would add a lot of clutter.
You end up with chains of functions, each taking 2 or 3 extra parameters
just because the function at the end of the chain needs them.