Sort kanji by stroke count and add option for shuffling the list
Previously the kanji list was in some order roughly (but not exactly)
corresponding to complexity and/or usage frequency. They are now sorted
by stroke count which is a more logical order to learn them in.

Curriculum learning can now be configured to either use the sorted list
or to shuffle the list before training.
The graph accuracy_comparison.png has been modified accordingly.