Major overhaul to pipeline.py. Make thread/process shutdown much cleaner.
I discovered that the way I was handling the multi-threaded and
multi-processing modes wasn't really optimal and was giving a false
measurement of the performance boost. This commit switches from using
an asynchronous circular buffer allowing multiple reads of the same
sample to a multiprocessing.Queue which allows for fairer comparisons.

The new data show that the multi-threaded version performs _worse_ than
the simple mode which I wrongly believed was single-threaded. In fact
TF already implements thread parallelism when building a Dataset from
a python generator function.

This commit also improves error handling and makes sure that all
background threads and processes exit cleanly even in the event of a
crash.